chore(project): fix source level 8

+++ engine/src/test/java/org/camunda/bpm/engine/test/cmmn/decisiontask/DmnDecisionTaskResultListenerTest.java
@@ -116,7 +116,7 @@ public class DmnDecisionTaskResultListenerTest extends CmmnProcessEngineTestCase
     assertEquals(1, results.size());
     DmnDecisionResultEntries firstOutput = results.get(0);
 
-    assertEquals(0, firstOutput.getFirstEntry());
+    assertEquals(0, (int) firstOutput.getFirstEntry());
     assertEquals(Variables.integerValue(0), firstOutput.getFirstEntryTyped());
   }
 
@@ -134,7 +134,7 @@ public class DmnDecisionTaskResultListenerTest extends CmmnProcessEngineTestCase
     assertEquals(1, results.size());
     DmnDecisionResultEntries firstOutput = results.get(0);
 
-    assertEquals(12, firstOutput.getFirstEntry());
+    assertEquals(12, (int) firstOutput.getFirstEntry());
     assertEquals(Variables.integerValue(12), firstOutput.getFirstEntryTyped());
   }
 
@@ -145,7 +145,7 @@ public class DmnDecisionTaskResultListenerTest extends CmmnProcessEngineTestCase
     assertEquals(1, results.size());
     DmnDecisionResultEntries firstOutput = results.get(0);
 
-    assertEquals(33, firstOutput.getFirstEntry());
+    assertEquals(33, (int) firstOutput.getFirstEntry());
     assertEquals(Variables.integerValue(33), firstOutput.getFirstEntryTyped());
   }
